Refactored: Move CI code to Datadog::CI namespace.
