+++ lib/datadog/ci.rb
@@ -1,0 +1,11 @@
+require 'ddtrace'
+
+module Datadog
+  # Namespace for Datadog CI instrumentation:
+  # e.g. rspec, cucumber, etc...
+  module CI
+  end
+end
+
+require 'datadog/ci/contrib/cucumber/integration'
+require 'datadog/ci/contrib/rspec/integration'
