+++ lib/datadog/ci/ext/app_types.rb
@@ -1,0 +1,9 @@
+module Datadog
+  module CI
+    module Ext
+      module AppTypes
+        TEST = 'test'.freeze
+      end
+    end
+  end
+end
