+++ lib/datadog/ci/test.rb
@@ -1,0 +1,8 @@
+module Datadog
+  module CI
+    # Common behavior for CI tests
+    module Test
+      # TODO: Extract some useful, common test behavior here.
+    end
+  end
+end
