+++ lib/ddtrace.rb
@@ -48,7 +48,6 @@ require 'ddtrace/contrib/active_record/integration'
 require 'ddtrace/contrib/active_support/integration'
 require 'ddtrace/contrib/aws/integration'
 require 'ddtrace/contrib/concurrent_ruby/integration'
-require 'ddtrace/contrib/cucumber/integration'
 require 'ddtrace/contrib/dalli/integration'
 require 'ddtrace/contrib/delayed_job/integration'
 require 'ddtrace/contrib/elasticsearch/integration'
@@ -75,10 +74,12 @@ require 'ddtrace/contrib/rake/integration'
 require 'ddtrace/contrib/redis/integration'
 require 'ddtrace/contrib/resque/integration'
 require 'ddtrace/contrib/rest_client/integration'
-require 'ddtrace/contrib/rspec/integration'
 require 'ddtrace/contrib/sequel/integration'
 require 'ddtrace/contrib/shoryuken/integration'
 require 'ddtrace/contrib/sidekiq/integration'
 require 'ddtrace/contrib/sinatra/integration'
 require 'ddtrace/contrib/sneakers/integration'
 require 'ddtrace/contrib/sucker_punch/integration'
+
+# Load extensions from CI namespace
+require 'datadog/ci'
