--- lib/ddtrace/contrib/cucumber/patcher.rb
@@ -1,23 +1,0 @@
-require 'ddtrace/contrib/patcher'
-require 'ddtrace/contrib/cucumber/instrumentation'
-
-module Datadog
-  module Contrib
-    module Cucumber
-      # Patcher enables patching of 'cucumber' module.
-      module Patcher
-        include Contrib::Patcher
-
-        module_function
-
-        def target_version
-          Integration.version
-        end
-
-        def patch
-          ::Cucumber::Runtime.include(Instrumentation)
-        end
-      end
-    end
-  end
-end
