+++ lib/ddtrace/ext/app_types.rb
@@ -6,7 +6,6 @@ module Datadog
       CACHE = 'cache'.freeze
       WORKER = 'worker'.freeze
       CUSTOM = 'custom'.freeze
-      TEST = 'test'.freeze
     end
   end
 end
