+++ spec/datadog/ci/contrib/cucumber/formatter_spec.rb
@@ -26,7 +26,7 @@ RSpec.describe 'Cucumber formatter' do
   end
 
   context 'executing a test suite' do
-    let(:args) { ['spec/ddtrace/contrib/cucumber/cucumber.features'] }
+    let(:args) { ['spec/datadog/ci/contrib/cucumber/cucumber.features'] }
 
     def do_execute
       cli.execute!(existing_runtime)
@@ -41,9 +41,9 @@ RSpec.describe 'Cucumber formatter' do
       step_span = spans.find { |s| s.resource == 'datadog' }
 
       expect(scenario_span.resource).to eq('cucumber scenario')
-      expect(scenario_span.service).to eq(Datadog::Contrib::Cucumber::Ext::SERVICE_NAME)
-      expect(scenario_span.span_type).to eq(Datadog::Ext::AppTypes::TEST)
-      expect(scenario_span.name).to eq(Datadog::Contrib::Cucumber::Ext::OPERATION_NAME)
+      expect(scenario_span.service).to eq(Datadog::CI::Contrib::Cucumber::Ext::SERVICE_NAME)
+      expect(scenario_span.span_type).to eq(Datadog::CI::Ext::AppTypes::TEST)
+      expect(scenario_span.name).to eq(Datadog::CI::Contrib::Cucumber::Ext::OPERATION_NAME)
       expect(step_span.resource).to eq('datadog')
     end
   end
