+++ spec/datadog/ci/contrib/cucumber/patcher_spec.rb
@@ -1,9 +1,9 @@
 require 'ddtrace/contrib/support/spec_helper'
-require 'ddtrace/contrib/cucumber/patcher'
+require 'datadog/ci/contrib/cucumber/patcher'
 
 require 'cucumber/runtime'
 
-RSpec.describe Datadog::Contrib::Cucumber::Patcher do
+RSpec.describe Datadog::CI::Contrib::Cucumber::Patcher do
   describe '.patch' do
     subject!(:patch) { described_class.patch }
