+++ spec/datadog/ci/ext/fixtures/ci/bitbucket.json
