+++ spec/datadog/ci/ext/fixtures/ci/bitrise.json
