+++ spec/datadog/ci/ext/fixtures/ci/gitlab.json
